refactor: refatorando controllers e usecases

+++ src/core/usecase/client/delete.ts
@@ -1,3 +1,7 @@
+import { IHttpResponse } from "../../../adapter/http-adapter/http-adapter"
+import { InvalidParamError, MissingParamError } from "../../../helper/errors"
+import { badRequest } from "../../../helper/errors/http-errors"
+import { ok } from "../../../helper/http-response"
 import { IClientRepository } from "../../repository/client"
 
 interface IRequest {
@@ -10,8 +14,19 @@ export class DeleteClientUsecase
         this.ClientRepository = ClientRepository
     }
 
-   async execute(request: IRequest): Promise<object> {
-        await this.ClientRepository.delete({id: request.id})
-        return {"message": `Deleted id ${request.id} with success.`}
+   async execute(request: IRequest): Promise<IHttpResponse> {
+       try {
+            if (!request.id) {
+                return badRequest(new MissingParamError('id'))
+            }
+            const idExists = await this.ClientRepository.getOne({id: request.id})
+            if (!idExists) {
+                return badRequest(new InvalidParamError(`ID[${request.id}] not found!`))
+            }
+            await this.ClientRepository.delete({id: request.id})
+            return ok(`Deleted id ${request.id} with success.`)
+       } catch (error) {
+            return error
+       }
     }
 }
